fix(model provision) rename files for more explicit name
